feat: rearchitecture and up \ cd
