+++ src/main/main.js
@@ -1,38 +1,47 @@
 import { argv, chdir, cwd, stdin, stdout } from 'node:process';
 import { EOL } from 'os';
-import { getName, finishWork, onStop } from './processes.js';
-import getFiles from './fs/getFiles.js';
-import { createStartPath, getBasePath } from './path.js';
-import { userAreInRootPath } from '../utils/textForUser.js';
+import { getName, finishWork, onStop } from '../modules/processes.js';
+import getFiles from '../modules/fs/getFiles.js';
+import { createStartPath, getPath } from '../modules/path/path.js';
+import up from '../modules/path/up.js';
+import { fail } from '../utils/constants.js';
 
 const main = async () => {
   createStartPath();
   getName(argv);
   stdin.on('data', async (data) => {
-    const input = data.toString().trim();
+    const input = data.toString().trim().split(' ');
+    const command = input[0];
+    console.log('input', input);
+    console.log('data', data.toString().trim());
     stdout.write(EOL);
-    switch (input) {
+    switch (command) {
       case 'exit':
       case '.exit':
         finishWork();
         break;
       case 'ls':
         const files = await getFiles(cwd());
-        console.log(`files:${EOL}`, files);
+        console.log(`files:${EOL}`, files, EOL);
         break;
       case 'up':
         try {
-          if (getBasePath(cwd()) === cwd()) {
-            userAreInRootPath();
-          } else {
-            chdir('../');
-          }
-          // process.chdir(env.userpath)
+          up();
         } catch (error) {
-          console.log(error);
+          console.error(fail);
         }
-        // я делал проверку path.isAbsolute, если true, то использую как есть, если false, то join и resolve
-        // updateDirectory = join(homeDirectory, ' ??' )
+        break;
+      case 'cd':
+        const path = input[1];
+        if (path) {
+          try {
+            chdir(path);
+          } catch (error) {
+            console.error(fail);
+          }
+        } else {
+          console.error(fail);
+        }
         break;
       default:
         stdout.write(`Invalid input${EOL}${EOL}`);
